# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.BindKnowledgeTagReq:
    properties:
      knowledgeId:
        type: string
      option:
        type: integer
      tagIdList:
        items:
          type: string
        type: array
    required:
    - knowledgeId
    - tagIdList
    type: object
...
  request.CreateKnowledgeTagReq:
    properties:
      tagName:
        type: string
    required:
    - tagName
    type: object
...
  request.DeleteKnowledgeTagReq:
    properties:
      tagId:
        type: string
    required:
    - tagId
    type: object
...
  request.KnowledgeTransferUserAdminReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUser:
        allOf:
        - $ref: '#/definitions/request.KnowledgeUserInfo'
        description: 知识库用户信息
      permissionId:
        description: 权限id编辑时传入
        type: string
    required:
    - knowledgeId
    - knowledgeUser
    type: object
  request.KnowledgeUserAddReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUserList:
        description: 知识库用户信息
        items:
          $ref: '#/definitions/request.KnowledgeUserInfo'
        type: array
      permissionType:
        description: '权限类型:0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑 授权权限>编辑权限>查看权限'
        type: integer
    required:
    - knowledgeId
    - knowledgeUserList
    type: object
  request.KnowledgeUserDeleteReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      permissionId:
        description: 知识库用户信息权限信息
        type: string
    required:
    - knowledgeId
    - permissionId
    type: object
  request.KnowledgeUserEditReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUser:
        allOf:
        - $ref: '#/definitions/request.KnowledgeUserInfo'
        description: 知识库用户信息
    required:
    - knowledgeId
    - knowledgeUser
    type: object
  request.KnowledgeUserInfo:
    properties:
      orgId:
        type: string
      permissionId:
        description: 权限id编辑时传入
        type: string
      permissionType:
        description: '权限类型: -1 删除此用户权限；0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑
          授权权限>编辑权限>查看权限'
        type: integer
      userId:
        type: string
    required:
    - orgId
    - userId
    type: object
...
  request.UpdateKnowledgeTagReq:
    properties:
      tagId:
        type: string
      tagName:
        type: string
    required:
    - tagId
    - tagName
    type: object
...
  response.CreateKnowledgeTagResp:
    properties:
      knowledgeId:
        type: string
    type: object
...
  response.KnowOrgInfo:
    properties:
      orgId:
        type: string
      orgName:
        type: string
    type: object
  response.KnowOrgUserInfoResp:
    properties:
      orgId:
        type: string
      orgName:
        type: string
      userInfoList:
        items:
          $ref: '#/definitions/response.KnowUserInfo'
        type: array
    type: object
  response.KnowUserInfo:
    properties:
      userId:
        type: string
      userName:
        type: string
    type: object
...
  response.KnowledgeTag:
    properties:
      selected:
        description: 此表标签是否选中
        type: boolean
      tagId:
        description: 知识库标签id
        type: string
      tagName:
        description: 知识库标签名称
        type: string
    type: object
  response.KnowledgeTagListResp:
    properties:
      knowledgeTagList:
        items:
          $ref: '#/definitions/response.KnowledgeTag'
        type: array
    type: object
  response.KnowledgeUserInfo:
    properties:
      orgId:
        type: string
      orgName:
        type: string
      permissionId:
        type: string
      permissionType:
        description: '权限类型: -1 删除此用户权限；0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑
          授权权限>编辑权限>查看权限'
        type: integer
      transfer:
        description: 是否显示转让按钮
        type: boolean
      userId:
        type: string
      userName:
        type: string
    type: object
  response.KnowledgeUserPermissionResp:
    properties:
      knowledgeUserInfoList:
        items:
          $ref: '#/definitions/response.KnowledgeUserInfo'
        type: array
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    ## HTTP Header
    | Header        | 说明      |
    | ------------- | --------- |
    | Authorization | JWT token |
    | X-Language    | 语言Code  |
    | X-Org-Id      | 组织ID    |
    | X-Client-Id   | 客户端标识|

    ## HTTP Status
    | HTTP Status             | 说明                   |
    | ----------------------- | ---------------------- |
    | 200, StatusOK           | 请求返回成功           |
    | 400, StatusBadRequest   | 请求返回失败，用于业务 |
    | 401, StatusUnauthorized | JWT认证失败            |
    | 403, StatusForbidden    | 没有权限               |

    ## 权限-菜单对应表
    | 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |
    |-------------|-------|------|------|------|------|
    | guest       |       |      | 【访客】 |      |      |
    | common      |       |      | 【通用】 |      |      |
    | permission  |       |      | 权限管理 |      |      |
    | permission  | user  |      | 权限管理 | 用户管理 |      |
    | permission  | org   |      | 权限管理 | 组织管理 |      |
    | permission  | role  |      | 权限管理 | 角色管理 |      |

    ## `/v1/user/permission`返回用例
    ```json
    {
      "code": 0,
      "data": {
        "orgPermission": {
          "org": {"id": "test-org-id", "name": "test-org-name"},
          "permissions": [
            {"perm": "permission"},
            {"perm": "permission.user"},
            {"perm": "permission.org"},
            {"perm": "permission.role"}
          ]
        }
      },
      "msg": "操作成功"
    }
    ```
...
paths:
...
  /knowledge/org:
    get:
      consumes:
      - application/json
      description: 知识库组织列表
      parameters:
      - description: 知识库id
        in: query
        name: knowledgeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowOrgInfo'
              type: object
      security:
      - JWT: []
      summary: 知识库组织列表
      tags:
      - knowledge
...
  /knowledge/tag:
    delete:
      consumes:
      - application/json
      description: 删除知识库标签
      parameters:
      - description: 删除知识库标签请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteKnowledgeTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除知识库标签
      tags:
      - knowledge.tag
    get:
      consumes:
      - application/json
      description: 查询知识库标签列表
      parameters:
      - in: query
        name: knowledgeId
        type: string
      - in: query
        name: tagName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeTagListResp'
              type: object
      security:
      - JWT: []
      summary: 查询知识库标签列表
      tags:
      - knowledge.tag
    post:
      consumes:
      - application/json
      description: 创建知识库标签
      parameters:
      - description: 创建知识库标签请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateKnowledgeTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CreateKnowledgeTagResp'
              type: object
      security:
      - JWT: []
      summary: 创建知识库标签
      tags:
      - knowledge.tag
    put:
      consumes:
      - application/json
      description: 修改知识库标签
      parameters:
      - description: 修改知识库标签请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateKnowledgeTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改知识库标签
      tags:
      - knowledge.tag
  /knowledge/tag/bind:
    post:
      consumes:
      - application/json
      description: 绑定知识库标签
      parameters:
      - description: 绑定知识库标签请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BindKnowledgeTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 绑定知识库标签
      tags:
      - knowledge.tag
  /knowledge/tag/bind/count:
    get:
      consumes:
      - application/json
      description: 查询标签绑定知识库数量
      parameters:
      - in: query
        name: tagId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeTagListResp'
              type: object
      security:
      - JWT: []
      summary: 查询标签绑定知识库数量
      tags:
      - knowledge.tag
  /knowledge/user:
    get:
      consumes:
      - application/json
      description: 知识库用户权限列表
      parameters:
      - description: 知识库id
        in: query
        name: knowledgeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeUserPermissionResp'
              type: object
      security:
      - JWT: []
      summary: 知识库用户权限列表
      tags:
      - knowledge
  /knowledge/user/add:
    post:
      consumes:
      - application/json
      description: 增加知识库用户
      parameters:
      - description: 增加知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 增加知识库用户
      tags:
      - knowledge
  /knowledge/user/admin/transfer:
    post:
      consumes:
      - application/json
      description: 转让管理员权限
      parameters:
      - description: 转让管理员权限请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeTransferUserAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 转让管理员权限
      tags:
      - knowledge
  /knowledge/user/delete:
    delete:
      consumes:
      - application/json
      description: 删除知识库用户
      parameters:
      - description: 删除知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除知识库用户
      tags:
      - knowledge
  /knowledge/user/edit:
    post:
      consumes:
      - application/json
      description: 修改知识库用户
      parameters:
      - description: 修改知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改知识库用户
      tags:
      - knowledge
  /knowledge/user/no/permit:
    get:
      consumes:
      - application/json
      description: 没有知识库用户列表
      parameters:
      - description: 知识库id
        in: query
        name: knowledgeId
        required: true
        type: string
      - description: 选择组织id
        in: query
        name: orgId
        required: true
        type: string
      - description: 是否是转让列表
        in: query
        name: transfer
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowOrgUserInfoResp'
              type: object
      security:
      - JWT: []
      summary: 没有知识库用户列表
      tags:
      - knowledge
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"